config/requirements.txt
python-gitlab==0.16
requests==2.11.1
six==1.10.0
